figured out the prob spot
Working through breaking points in console I realized that the problem is that once the first graph data is processed (filtering the dataset down to two properties) the input dataset is not redefined to give the second prepData() the full set of variables.
I tried redefining thisData to lineData at the beginning of the _.each() process but when prepData() executes the first graph's data, both lineData and thisData changes.