map hover effect achieved
when mouse hovers over a point, the cursor changes to pointer style, and the circle marker gets 3x bigger and changes color to purple.
Struggled with the layer hierarchy. As far as i can tell, whatever is placed first is below what comes next. So the hover effect layer should come after the original layer.